Lint code

+++ part4/blog-list/controllers/blogs.js
@@ -5,56 +5,70 @@
 */
 
 const blogsRouter = require('express').Router();
+const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
-const jwt = require('jsonwebtoken');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({})
     .populate('user', { username: 1, name: 1 });
 
-  response.json(blogs);
+  return response.json(blogs);
 });
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id);
 
   if (blog) {
-    response.status(200).json(blog);
-  } else {
-    response.status(404).end();
+    return response.status(200).json(blog);
   }
+  return response.status(404).end();
 });
 
 blogsRouter.post('/', async (request, response) => {
   const body = new Blog(request.body);
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
 
   if (!decodedToken.id) {
-    return response.status(401).json({ error: 'invalid token' })
+    return response.status(401).json({ error: 'invalid token' });
   }
-  const user = await User.findById(decodedToken.id);
+
+  const user = request.user
+  // const user = await User.findById(decodedToken.id);
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    user: user.id
+    user: user.id,
   });
 
   const savedBlog = await blog.save();
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save();
 
-  response.status(201).json(savedBlog);
+  return response.status(201).json(savedBlog);
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+  const user = request.user
+  // const user = await User.findById(decodedToken.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    return response.status(204).end();
+  }
+  return response.status(400).json({ error: 'invalid user' });
 });
 
 blogsRouter.put('/:id', (request, response, next) => {
